ZFIN-7932:
Make the output easier to interpret on failure.
Add check for existing file.

+++ server_apps/DB_maintenance/Alliance/validateAllianceFiles.sh
@@ -19,6 +19,18 @@ main() {
   validateFile ZFIN_1.0.1.4_Resource.json.gz RESOURCE
   validateFile ZFIN_1.0.1.4_ReferenceExchange.json.gz REF-EXCHANGE
 
+  echo ""
+
+  #Display error if found
+  if [ $BUILD_STATUS_CODE -ne 0 ]; then
+    echo "---------------"
+    echo "FAILED UPLOADS:"
+    echo "---------------"
+    echo "$FAILED_UPLOADS" | sed 's/, $//' #remove trailing comma
+  fi
+
+  echo ""
+
   exit $BUILD_STATUS_CODE
 }
 
@@ -37,6 +49,7 @@ loadConfig() {
 
   RELEASE_VERSION=$2
   BUILD_STATUS_CODE=0
+  FAILED_UPLOADS=""
   ENDPOINT=validate
   if [ "$1" == "true" ]; then
         ENDPOINT=submit
@@ -57,23 +70,37 @@ validateFile() {
   FIELD_NAME=$2
   FRIENDLY_FILENAME=$JSON_FILENAME
   TEMP_RESPONSE_FILE=/tmp/agr_upload_response.txt
+  FILE_EXISTS="true"
 
-  #Validate file (or submit)
-  echo ""
-  echo "Validating $FRIENDLY_FILENAME file..."
-  echo "curl --silent -H \"Authorization: Bearer AUTHORIZATION\" -X POST \"$BASE_URL/api/data/$ENDPOINT\" -F \"${RELEASE_VERSION}_${FIELD_NAME}_ZFIN=@${JSON_FILENAME}\""
-  curl --silent -H "Authorization: Bearer $AUTHORIZATION" -X POST "$BASE_URL/api/data/$ENDPOINT" -F "${RELEASE_VERSION}_${FIELD_NAME}_ZFIN=@${JSON_FILENAME}" | tee $TEMP_RESPONSE_FILE
+  #check if validation file exists
+  ls "$JSON_FILENAME" > /dev/null
+  if [ $? -ne 0 ]; then
+    echo "ERROR: File not found: '$JSON_FILENAME'"
+    FILE_EXISTS="false"
+    EXIT_CODE=1
+  fi
 
-  #Check server response for failure
-  #If server set response code to an error status code in the event of status:failed, we could get curl exit code, but it currently sends back a 200
-  grep -qv '"status":"failed"' $TEMP_RESPONSE_FILE
-  EXIT_CODE=$?
-  rm $TEMP_RESPONSE_FILE
+  if [ $FILE_EXISTS == "true" ]; then
+    #Validate file (or submit)
+    echo ""
+    echo "Validating $FRIENDLY_FILENAME file..."
+    echo "curl --silent -H \"Authorization: Bearer AUTHORIZATION\" -X POST \"$BASE_URL/api/data/$ENDPOINT\" -F \"${RELEASE_VERSION}_${FIELD_NAME}_ZFIN=@${JSON_FILENAME}\""
+    curl --silent -H "Authorization: Bearer $AUTHORIZATION" -X POST "$BASE_URL/api/data/$ENDPOINT" -F "${RELEASE_VERSION}_${FIELD_NAME}_ZFIN=@${JSON_FILENAME}" | tee $TEMP_RESPONSE_FILE
+
+    #Check server response for failure
+    #If server set response code to an error status code in the event of status:failed, we could get curl exit code, but it currently sends back a 200
+    grep -qv '"status":"failed"' $TEMP_RESPONSE_FILE
+    EXIT_CODE=$?
+    rm $TEMP_RESPONSE_FILE
+
+    echo ""
+  fi
 
   #Handle error if found
   if [ $EXIT_CODE -ne 0 ]; then
-      echo "ERROR: Failure response from server on $FRIENDLY_FILENAME file"
+      echo "ERROR: Failure validating $FRIENDLY_FILENAME file"
       BUILD_STATUS_CODE=$EXIT_CODE
+      FAILED_UPLOADS="$FRIENDLY_FILENAME, $FAILED_UPLOADS"
   fi
 
 }
